feat(cli): clean-up and refactor entrypoints (#75)
https://coveord.atlassian.net/browse/CDX-119
