+++ .vscode/launch.json
@@ -8,19 +8,7 @@
       "type": "pwa-node",
       "request": "attach",
       "cwd": "${workspaceFolder}",
-      "name": "Attach",
-      "address": "localhost",
-      "port": 9229,
-      "localRoot": "${workspaceFolder}",
-      "remoteRoot": "/home/notGroot/cli-copy",
-      "sourceMaps": true,
-      "autoAttachChildProcesses": false
-    },
-    {
-      "type": "pwa-node",
-      "request": "attach",
-      "cwd": "${workspaceFolder}",
-      "name": "Brace",
+      "name": "Debug E2E test (Start)",
       "address": "localhost",
       "port": 9229,
       "localRoot": "${workspaceFolder}",
@@ -29,6 +17,18 @@
       "autoAttachChildProcesses": false,
       "preLaunchTask": "TestDaemon",
       "postDebugTask": "DaemonKiller"
+    },
+    {
+      "type": "pwa-node",
+      "request": "attach",
+      "cwd": "${workspaceFolder}",
+      "name": "Debug E2E test (Attach)",
+      "address": "localhost",
+      "port": 9229,
+      "localRoot": "${workspaceFolder}",
+      "remoteRoot": "/home/notGroot/cli-copy",
+      "sourceMaps": true,
+      "autoAttachChildProcesses": false
     }
   ]
 }
