+++ packages/cli-e2e/entrypoints/vsCodeExit.js
@@ -1,6 +1,4 @@
 const {execSync} = require('child_process');
+const {DOCKER_CONTAINER_NAME} = require('./common');
 
-const DOCKER_CONTAINER_NAME = 'coveo-cli-e2e-container';
-console.log('Doom guy is here');
-execSync(`docker container rm ${DOCKER_CONTAINER_NAME}`);
-console.log('Daemon BFGed');
+execSync(`docker container rm ${DOCKER_CONTAINER_NAME} -f`);
